requirements.txt
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-compressor==4.4
django-cors-headers==4.2.0
django-ninja==0.22.2
Pillow==10.0.0
pydantic==1.10.12
PyJWT==2.8.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
typing_extensions==4.7.1
